sql: make interval planning across WITH options and expressions
WITH option planning previously ignored interval modifiers; at variance
from how intervals were planned when used in expressions. This commit
refactors the interval planning code into a helper method that is shared
by both expression planning and WITH option planning.

This commit also introduces a new `TryFromValue` implementation
targeting `Interval`, to tee up the planning of the forthcoming
`REFRESH` option (#23870), which wants the interval as an `Interval`
rather than as a `Duration`.
